Implement the hdr-metadata-unstable-v1 protocol
This is just a placeholder to get data from clients to test our HDR
stack. The actual protocol is being discussed here:
https://patchwork.freedesktop.org/series/57657/
The protocol is to set the colorspace details of the surface. The HDR
metadata would also be part of that protocol.

v2: Use the common color primary definitions (Pekka)

Signed-off-by: Harish Krupo <harish.krupo.kps@intel.com>
Signed-off-by: alex.chae <alex.chae@intel.com>