compositor-drm: Prepare connector's color state
This patch adds initial code to handle a HDR view in DRM backend,
while plane assignment. In order to handle HDR playback, the first
step is to identify a HDR view, by identifying a HDR surface. ie,
if any of the surface comes with HDR metadata, its a HDR surface,
which makes this view a HDR view.

Now, once we identify an HDR view, we have to set output colorspace
If output is HDR -> set output colorspace as REC2020/DCIP3
If output is SDR -> set output colorspace as REC709/default

This patch prepares the required calculations, and prepares the
connector's color state to be commited, and reflect these changes
in current atomic commit.

V2: Addressed Pekka's review comments

Signed-off-by: Shashank Sharma <shashank.sharma@intel.com>
Signed-off-by: alex.chae <alex.chae@intel.com>