looking-glass: Added a remote plugin for looking-glass (v2)
This plugin allows for the Weston compositor to provide access of its
framebuffer to a Looking Glass Wayland client app. The app can
communicate via the lg-remote interface. The interface looks quite
similar to screenshot one where the app gets one request to capture and
one event when it is done. The app needs to allocate a SHM buffer which
it provides to the compositor to fill up. This buffer needs to be at
least the size of the frame buffer of the output for which the
compositor is capturing. The compositor waits for the rendering to
finish by the GPU and then copies the buffer from DRM prime handle into
the SHM buffer provided by the app. When done, it sends the done event
to the app notifying that it is now safe to use that buffer.
lg-sample app is just a sample showing how to get this frame buffer and
communicate via the lg-remote interface. The interface looks quite
similar to screenshot one where the app gets one request to capture and
one event when it is done. The app needs to allocate a SHM buffer which
it provides to the compositor to fill up. This buffer needs to be at
least the size of the frame buffer of the output for which the
compositor is capturing. The compositor waits for the rendering to
finish by the GPU and then copies the buffer from DRM prime handle into
the SHM buffer provided by the app. When done, it sends the done event
to the app notifying that it is now safe to use that buffer.
lg-sample app is just a sample showing how to get this frame buffer and
write it into a PNG using cairo lib.

v2:

Since lg-remoting was started off as a copy of the remoting plugin,
there were a lot of remnants from the old gstreamer code. With this
change, we have removed all of those.
There was also a statement where the size of the FB was hardcoded. This
is now calculated based off the width and height.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>